pyproject.toml
[tool.poetry]
name = "ms-inventario"
version = "0.1.0"
description = "Microservicio FastAPI en Python 3.13 para Cloud Run"
authors = ["G17 <team@example.com>"]
readme = "README.md"
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = "^3.13"
fastapi = ">=0.110"
uvicorn = {extras=["standard"], version=">=0.29"}
pydantic = ">=2.7"
pydantic-settings = ">=2.2"
sqlalchemy = {extras=["asyncio"], version=">=2.0"}
asyncpg = ">=0.29"
redis = {extras=["async"], version=">=5.0"}
google-cloud-pubsub = ">=2.21"
google-cloud-bigquery = ">=3.25"
google-cloud-bigtable = ">=2.24"
google-cloud-storage = ">=2.18"

[tool.poetry.group.dev.dependencies]
pytest = ">=8.2"
httpx = ">=0.27"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
